refactor: extract ComposeFilesController from ComposeController (PR6, seam 2)
Second seam of the ComposeController split. Moves the compose discovery / health
endpoints into a focused ComposeFilesController (explicit [Route("api/compose")],
public routes unchanged):

  GET  files
  GET  conflicts
  GET  health      ([AllowAnonymous])
  POST refresh     (admin)

Removes three now-unused dependencies from ComposeController
(IConflictResolutionService, IOptions<ComposeDiscoveryOptions>, DockerService);
IComposeFileCacheService stays (still used by the projects refresh path).

Adds ComposeFilesControllerTests (5 tests: file mapping, self-project filtering,
conflicts, refresh auth + invalidate/scan/audit). The health endpoint depends on
the live DockerService and is covered by the app-level /health smoke instead.

ComposeController: 1631 -> 1418 lines.

Verify: dotnet build -c Release -> 0 warnings/errors; dotnet test -> 320 passed;
app boots, routes resolve (files/conflicts/refresh 401, health 200, no ambiguous
route).
